# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/tenants": {
      "post": {
        "operationId": "TenancyController_register",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated ULID key; replays return the original response",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 26,
              "maxLength": 26,
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRegistrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key, or invalid body",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Owner email already registered (duplicate-email)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Idempotency key reused with a different payload (idempotency-key-reuse)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Registers a tenant with its Owner user",
        "tags": [
          "tenancy"
        ]
      }
    },
...
    "/tenants/{tenantId}/users": {
      "post": {
        "operationId": "UsersController_inviteUser",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key, or invalid body",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied), or the caller lacks users.invite (role-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already has an account in any tenant (email-exists), or a concurrent idempotent request (conflict)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Idempotency key reused with a different payload (idempotency-key-reuse)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invites a user (Owner only) — returns the one-time invite token",
        "tags": [
          "tenancy"
        ]
      },
      "get": {
        "operationId": "UsersController_listUsers",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque keyset cursor from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed cursor (invalid-cursor) or out-of-range limit (validation-failed)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lists the tenant’s users (keyset cursor pagination — open to any member)",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenants/{tenantId}/users/{userId}": {
      "patch": {
        "operationId": "UsersController_setUserRole",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key, or invalid body",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied), or the caller lacks users.role_change (role-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "User does not exist in this tenant (not-found)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "The target is the tenant’s last Owner (last-owner), or a concurrent idempotent request (conflict)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changes a user’s role (Owner capability users.role_change; effective on the user’s next command)",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenants/{tenantId}/accept-invite": {
      "post": {
        "operationId": "UsersController_acceptInvite",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "The inviting tenant (embedded in the invite link)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated ULID key; replays return the original response",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 26,
              "maxLength": 26,
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInviteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed Idempotency-Key, invalid body, or an unknown/used/expired token (invite-invalid)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Idempotency key reused with a different payload (idempotency-key-reuse)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Accepts a one-time invite (unauthenticated): sets the account password and activates the user",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenants/{tenantId}/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Owning tenant (must match the session)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Session belongs to another tenant (permission-denied)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The signed-in user’s own row (id, email, role, status)",
        "tags": [
          "tenancy"
        ]
      }
    },
...
  },
  "info": {
    "title": "WMS API",
    "description": "Warehouse Management System backend — modular monolith. This document is the contract: web and mobile clients generate typed clients from it (AD-8).",
    "version": "1.0.0",
    "contact": {
      "name": "WMS",
      "url": "https://wms.example.com",
      "email": "api@wms.example.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wms.example.com/license"
    }
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://wms.example.com/problems/validation-failed"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "validation-failed"
          },
          "detail": {
            "type": "string",
            "example": "Echo body must be a JSON object"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/echo"
          },
          "errors": {
            "example": [
              "key should not exist"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ]
      },
      "RegisterTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Priya Spices Pvt Ltd",
            "minLength": 1,
            "maxLength": 200
          },
          "ownerEmail": {
            "type": "string",
            "example": "priya@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery",
            "minLength": 8,
            "maxLength": 200
          }
        },
        "required": [
          "name",
          "ownerEmail",
          "password"
        ]
      },
      "TenantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "0198f7a2-1b3c-7d4e-8f90-112233445566"
          },
          "name": {
            "type": "string",
            "example": "Priya Spices Pvt Ltd"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "OwnerUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "0198f7a2-1b3c-7d4e-8f90-112233445599"
          },
          "email": {
            "type": "string",
            "example": "priya@example.com"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "TenantRegistrationResponse": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantResponse"
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerUserResponse"
          }
        },
        "required": [
          "tenant",
          "owner"
        ]
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "dev@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "ops_manager",
              "operator",
              "accountant"
            ],
            "example": "owner"
          },
          "status": {
            "type": "string",
            "enum": [
              "invited",
              "active"
            ],
            "example": "active"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-08T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status",
          "createdAt"
        ]
      },
...
      "InviteUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "inviteToken": {
            "type": "string",
            "description": "The one-time invite token — share the accept link out-of-band"
          },
          "inviteExpiresAt": {
            "type": "string",
            "example": "2026-09-15T00:00:00.000Z",
            "description": "One-time link expiry (7 days)"
          }
        },
        "required": [
          "user",
          "inviteToken",
          "inviteExpiresAt"
        ]
      },
      "UserListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Opaque keyset cursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "AcceptInviteDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The one-time invite token from the invite response"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery",
            "minLength": 8,
            "maxLength": 200
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "AcceptInviteResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user"
        ]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user"
        ]
      },
...
    }
  }
}